day_89/requirements.txt
streamlit==1.32.0
ultralytics==8.1.0
Pillow==10.2.0
opencv-python==4.9.0.80
gtts==2.5.1
eng-to-ipa==0.0.2
pandas==2.1.4
plotly==5.18.0
numpy==1.24.3 